[ISSUE-805][REFACTOR] Remove UserIdentifier out of ControlMessage
