+++ client/src/main/java/org/apache/celeborn/client/ShuffleClient.java
@@ -24,7 +24,7 @@ import org.apache.hadoop.fs.FileSystem;
 
 import org.apache.celeborn.client.read.RssInputStream;
 import org.apache.celeborn.common.RssConf;
-import org.apache.celeborn.common.protocol.message.ControlMessages.UserIdentifier;
+import org.apache.celeborn.common.identity.UserIdentifier;
 import org.apache.celeborn.common.rpc.RpcEndpointRef;
 
 /**
