+++ common/src/main/java/org/apache/celeborn/common/meta/FileInfo.java
@@ -27,7 +27,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 
-import org.apache.celeborn.common.protocol.message.ControlMessages.UserIdentifier;
+import org.apache.celeborn.common.identity.UserIdentifier;
 import org.apache.celeborn.common.util.Utils;
 
 public class FileInfo {
