+++ common/src/main/java/org/apache/celeborn/common/util/PbSerDeUtils.java
@@ -22,6 +22,8 @@ import java.util.concurrent.ConcurrentHashMap;
 
 import com.google.protobuf.InvalidProtocolBufferException;
 
+import org.apache.celeborn.common.identity.UserIdentifier;
+import org.apache.celeborn.common.identity.UserIdentifier$;
 import org.apache.celeborn.common.meta.DiskInfo;
 import org.apache.celeborn.common.meta.FileInfo;
 import org.apache.celeborn.common.meta.WorkerInfo;
