+++ common/src/test/scala/org/apache/celeborn/common/identity/UserIdentifierSuite.scala
@@ -19,7 +19,6 @@ package org.apache.celeborn.common.identity
 
 import org.apache.celeborn.RssFunSuite
 import org.apache.celeborn.common.exception.RssException
-import org.apache.celeborn.common.protocol.message.ControlMessages.UserIdentifier
 
 class UserIdentifierSuite extends RssFunSuite {
