+++ worker/src/test/scala/org/apache/celeborn/service/deploy/cluster/ClusterReadWriteSuite.scala
@@ -27,7 +27,7 @@ import org.scalatest.funsuite.AnyFunSuite
 import org.apache.celeborn.client.{LifecycleManager, ShuffleClientImpl}
 import org.apache.celeborn.client.compress.Compressor.CompressionCodec
 import org.apache.celeborn.common.RssConf
-import org.apache.celeborn.common.protocol.message.ControlMessages.UserIdentifier
+import org.apache.celeborn.common.identity.UserIdentifier
 import org.apache.celeborn.service.deploy.MiniClusterFeature
 
 class ClusterReadWriteSuite extends AnyFunSuite with MiniClusterFeature with BeforeAndAfterAll {
